Put each board config in own file.
